requirements.txt
altair==4.2.0
appdirs==1.4.4
attrs==22.2.0
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cryptography==38.0.4
decorator==5.1.1
entrypoints==0.4
frozendict==2.3.4
gitdb==4.0.10
GitPython==3.1.29
html5lib==1.1
idna==3.4
importlib-metadata==5.2.0
investpy==1.0.8
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.2
MarkupSafe==2.1.1
multitasking==0.0.11
numpy==1.24.0
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
plotly==5.11.0
protobuf==3.20.3
pyarrow==10.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
streamlit==1.16.0
streamlit-toggle-switch==1.0.2
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
Unidecode==1.3.6
urllib3==1.26.13
validators==0.20.0
webencodings==0.5.1
yfinance==0.2.3
zipp==3.11.0
